requirements.txt
aldryn-apphooks-config==0.4.1
aldryn-boilerplates==0.7.7
aldryn-categories==1.1.0
aldryn-common==1.0.4
-e git+https://github.com/aldryn/aldryn-events.git@71eba670a5e8af7e845735de83e5bbb8edc38b49#egg=aldryn_events
aldryn-forms==3.0.3
aldryn-newsblog==2.1.1
aldryn-people==2.1.0
aldryn-translation-tools==0.2.1
aldryn-video==1.0.0
aldryn-wow==1.0.0
asn1crypto==0.24.0
Babel==2.6.0
backport-collections==0.1
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
cssselect==1.0.3
cssutils==1.0.2
dj-database-url==0.5.0
Django==1.11.14
django-absolute==0.3
django-appconf==1.0.2
django-appdata==0.2.0
django-bootstrap3==10.0.1
django-classy-tags==0.8.0
django-cms==3.5.2
django-debug-toolbar==1.9.1
django-emailit==0.2.3
django-extended-choices==1.3
django-filer==1.3.2
django-formtools==2.1
django-mptt==0.8.7
django-parler==1.9.2
django-phonenumber-field==1.3.0
django-polymorphic==2.0.2
django-ranged-response==0.2.0
django-sekizai==0.10.0
django-simple-captcha==0.5.9
django-sizefield==0.9.1
django-sortedm2m==1.5.0
django-standard-form==1.1.1
django-tablib==3.2
django-taggit==0.22.2
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-bootstrap4==1.0.0
djangocms-column==1.8.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-history==0.5.3
djangocms-icon==1.0.0
djangocms-inherit==0.2.2
djangocms-installer==1.0.1
djangocms-light-gallery==0.0.7
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-unitegallery==0.1.0
djangocms-video==2.0.4
easy-thumbnails==2.5
et-xmlfile==1.0.1
ez-setup==0.9
future==0.16.0
html5lib==0.9999999
idna==2.7
jdcal==1.4
jsonfield==2.0.2
lxml==4.2.3
micawber==0.3.5
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.4.9
phonenumbers==8.9.9
phonenumberslite==8.9.9
Pillow==5.2.0
pkg-resources==0.0.0
premailer==3.2.0
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
python-slugify==1.2.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
tablib==0.12.1
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
YURL==0.13
